Adding multi mesh import from same file
I've edited importMesh and isDescendantOf method to accepts array of meshes to import from same file. If you use .incremental file with this method it loads meshes one by one (like when you import whole .incremental file). I don't know if it's the clearest way to do that, but the idea is here ;)

Former-commit-id: 4132e9ee5b3875b3e930a4c4fd1e9358bf836ecb