Added a system that allows cameras to force a wireframe display mode when moving if the FPS falls down under a configurable value (can prevent some very laggy moments).
Totally optional and configurable. Added by default only 1 comparison and 1 assignation (each frame/each camera). All cameras will be able to force wireframe.

Use the camera.fpsThresoldForForceWireframe property to set the FPS value under which the wireframe mode will be forced when the camera is moving (set -1 to never force wireframe when the camera is moving and a very huge value : 100+ to always force it).