ArcRotateCamera, Mesh and Vector3 new methods and modified OurDemoController of BuildOurOwnBabylonJS
Added zoomOn and focusOn methods to BABYLON.ArcRotateCamera: it takes an array of meshes of parameters (can be undefined to use all scene meshes) on which the camera must zoomOn or only focusOn.
Added MinMax static method to BABYLON.Mesh: it takes an array of meshes of parameters (can be undefined to use all scene meshes) and retrieves the min and max vectors.
Added Center static method to BABYLON.Mesh: it takes an array of meshes of parameters (can be undefined to use all scene meshes) and retrieves the center of the bounding box wrapping all meshes.
Added MinimizeInPlace and MaximizeInPlace methods to BABYLON.Vector3 in the aim to speed up BABYLON.Mesh.MinMax
Added Center static method to BABYLON.Vector3

Now, the Show action of OurDemoController takes two arguments: 'viewName' and 'folder'. 'folder' is the path relative to 'Views/OurDemo' and 'viewName' is the name of the view in this folder. 'folder' is passed to the view so it can be used to retrieve the babylonJS file for example (see sample.cshtml)
Added sample.cshtml to quickly test babylonJS files against your modified version of BabylonJS