Fix for wrong intersection result
Preventing selection of a face that is behind the ray's origin position.
The problem is that if the origin point is INSIDE a mesh, the faces behind the point will be selected as well. The Distance received from the PickingInfo will still be positive, due to the new initialization in the scene pick function.

Former-commit-id: 8494252f3fef010942289d61b550afdab11dd1e3