Mesh Simplification porting to babylon JS
The mesh simplification is now a part of the framework. A
meshSimplification file was added, including an ISimplifier interface
and a SimplificationType enum that will allow further implementations of
simplifications.
Mesh now has a function "simplify" that receives a list of
SimplificationSettings (containing quality and distance for LOD) and add
them to the LOD list of the mesh.
The implementation allows using the simplification outside the mesh, if
needed somewhere else (it can be used to async clone a mesh, if quality
is set to 1.0, for example).
First commit, hopefully no typos or errors. Was tested locally.


Former-commit-id: 769e5f0e7b2fe802f81346e220c6b3411eaae2e2