Fix for imposter and options init
Since both are optional in typescript, they should be initialized in the
function before used.
importer was accessed before its initialization, moved the init a few
lines up.
options is optional but is accessed without checking if it exists. It is
now being initialized or set in case some values are missing.


Former-commit-id: 427bbff292a24d42a4597876f8640540f08d8f26