Fixing wrong filenames
Some of the filenames in the gulpfile are wrong.
A few files were moved from one directory to the other, the textures
directory was written in lowercase (and not in uppercase). Windows
ignores this, but linux is case sensitive.
This should somehow be automated. I will see if the new gulp-typescript
has filters for that. Shouldn't be a problem to automate js
file-finding.


Former-commit-id: 77d4bd925b3c9eb9925a4860a3a8944f15d7429a