Allow Clones to be merged VertexData.ExtractFrom() is called during the merge process.  It calls the geometry functions getVerticesData() & getIndices().  These just returned reference variables.
The call argument, copyWhenShared? : boolean, was added to these
functions.  It was also added to the VertexData extraction functions,
and those in Mesh as well.  When true, a copy is returned.

In Mesh.MergeMeshes this is set to true in the call to Extract.  After
the first Mesh, this unfortunately must still be true, since the
positions & normals are also transformed, requiring a private copy.

Mesh.MergeMeshes was also improved, allowing meshes array to contain
null items, which are just skipped.  An additional optional argument was
added, the mesh to apply the vertex data.  This allows the sub-classes
of BABYLON.Mesh to contain the result of the merge. 

Former-commit-id: bae7e6950d0f72cd47bcf831faa57ebf7ee314c5