Worker can be in a single file
If single files are loaded, the worked will load the needed scripts.
Worker's location has to be configured in line 214 of
babylon.collisionCoordinator.ts , as the path of the current script's
location is not a standard JS parameter.
The worker's js needs to be included in the main html file as well.
Legacy collision is not the default.
Gulp config was changed to fit the new needs.


Former-commit-id: df0a1bfc93944e66b9f90fa8b2eff4ecca6ae3a1