Fix for collision worker and gravity
There was a problem with parallel processing of gravity-enabled cameras.
The position that was used for normal collision and gravity collision
was the same, and it caused major problems, especially in IE.
This fix make the inspection a bit more "sync" - first the regular
inspection is made, and a temporary position buffer is being used to
inspect the gravity. This actually means that gravity inspection will
take a minimum of 1 frame (probably 2), but the movement will be
correct.


Former-commit-id: 1a43b95591b65061d8dfedef707da263625b3a40