LookAt now works in global space as well
The target point should be in the same space as the current object, but
sometimes it is not possible (for example, when you want to look at a
mesh that is a child of a different node).
If space is local (default) the position of the mesh is users, but if
space is world, the absolute position of the mesh is used, and then it
is expected that the targetPoint will also be in world space.
Also - made optional parameters really optional :smile: