Adjusting the WebVR architecture to fit the new WebVR API
The engine is now responsible to search for new devices, if WebVR is
enabled in the browser.
If WebVR is enabled (and a device was chosen (attachControl), the
fullscreen and request animation frame will be triggered by the
vrDisplay device.
Users with no WebVR enabled can still use the API in the same way.


Former-commit-id: 9e64b9c6513aeec3580cf6d205d1a103dd508096