Adjusting to the WebVR specs
Thou working when called multiple times, the requestAnimationFrame
should only be called once when the device is presenting.
Also added the possibility of auto-enabling WebVR, while not being
initialized automatically.  This was done because when WebVR is
initializing it starts the native VR software, which might not be
desired.
The WebVR camera forces WebVR init.