Support loading with a module loader.
Babylon.js currently uses namespaces and its types can be pulled in using

    /// <reference path="babylonjs/babylon.d.ts" />

It's generated JavaScript can be loaded using either a script tag, or using a module loader from JavaScript.
However, what was not yet supported is loading types and implementation using a module loader from Typescript,
because babylon.d.ts was not a module.

This adds support for that, following the example here:
https://github.com/Microsoft/TypeScript/wiki/What's-new-in-TypeScript#support-for-umd-module-definitions

Former-commit-id: 5f3e00bbc83074d1bb5f4e95d97d64c40c486854