Bug fixed, refereces changed
If pointerInput was already defined, the engine variable (local in the
attachControl scope) will be undefined.
I also changed all references of camera to this.camera to prevent
confusion.
The engine variable can also be referenced (this.camera.getEngine()),
but I am not sure about the getter's impact on performance.