Update babylon.abstractMesh.ts
I add a function getHeightMeshAtCoordinates, because when a terrain is serialize and reload with ImportMesh, getHeightAtCoordinates that is made for when creating a field with CreatGround, no longer works.
In this case, we can use this function which will return the point Y of a coordinate.

I also add setUpdatable on the mesh, because you can only put an object updatable to the creation of the object, but once serialize and reload with ImportMesh, it is no longer updatable. With this function you can redefine updatable.