WaterMaterial DeltaTime fix
Due to the way the water material handles the engine's delta time, it
was displayed wrong in VR.
The water material will now check if the last delta time is equal to the
delta time cached and will only add if it doesn't match. This will
prevent changing the internal delta time when rendering the 2nd eye.