Default Values for CreateDashedLines Added
If  options not set for gapSize and dashSize then variables on lines 425, and 426 are undefined and so undefined is stored in the mesh properties of gapSize and dashSize for dashedLines object on lines 477 and 478. This means that when an instance is used dashshft on line 445 is calculated using undefined variables resulting in NaN for positions[p + 3], positions[p + 4], positions[p + 5] on lines 456 to 458.

Former-commit-id: 7b76a8926c7d7daf7c434e347b8c47a65c6c0760