Add RedefineSerializedUrl for the Textures
Allows you to redefine a URl for the search path for images of serialized models that are in a path different from the original models.

If the path is redefine, UseSerializedUrlIfAny will be ignored if it has been set to true.

Why that :
On my project, I have on one side the editor that creates scenes and saves by serializing the models in a folder '_Projets/TitleGame/ scenes/'. The basic models are on another path. On the editor, no problem my models are reloading properly thanks to UseSerializedUrlIfAny.

Now if I run the game, I run it from TitleGame / search for images that made from '_Projects/TitleGame/' and suddenly no images of my models have the correct path.

RedefineSerializedUrl may allow me to redefine the correct path quite easily with this property.