Check length of XR input sources
If length is not checked, then if the `XRSession` returns a valid empty array for `inputSources`, a new event listener will be added to the session manager every XR frame. According to [the spec](https://immersive-web.github.io/webxr/#list-of-active-xr-input-sources), it appears that this valid-but-empty scenario can (and perhaps must) occur.

Former-commit-id: a1209c58ebb9f12fc067e4dcc14d1fd680a8484b