Initial implementation of NativeInput plugin (#234)
This is an initial implementation of NativeInput. With this change, it only supports touch input (we can add more later, but this puts the basic infrastructure in place). The main layers are:

1. NativeInput is the public native contract. It only exposes the ability to write input events (e.g. forward them from the platform specific APIs).
2. NativeInput::Impl is the internal contract for both writing input events, and reading (polling) input state.
3. DeviceInputSystem is the native "polyfill" (replacement) for the Babylon.js DeviceInputSystem. Instances of this NAPI based class simply read input state from the NativeInput::Impl instance.

Note that currently, a NativeInput instance is scoped to a Napi::Env. In the future (as noted in the comments), these should be scoped to individual views that the user can interact with.