Add Activity related events and support pause/resume in ARCore XR (#247)
This change adds three "events" to the Android globals: Pause, Resume, and RequestPermissionsResult. The first two are currently being used to pause/resume the ARCore session, and address one of the items in #230. I added the third because it was very similar to code I was already writing and @Alex-MSFT will need it for the permissions request stuff he is working on. I added a helper Event class, which could be moved to arcana, but I'm not sure if it will be used in the long term or not based on the AppLifecycle Core component we've discussed potentially adding, so I could go either way (leave it or move it to arcana).