Move app context jobject lifetime management to Globals (#252)
This change makes it the responsibility of Globals to manage the lifetime of the jobject referring to the app context. With the current implementation, it does this by first getting the application context from an arbitrary context, and creating a global ref from this jobject. The application context itself has the lifetime of the application, so in this case Globals just never cleans up this jobject (it will always exist, and reference an always existing application context). This addresses part of #230.