Add Function to retrieve all affected meshes
Rather than keeping an outside list of affected meshes by the HighlightLayer, we can easily return the private variable for iteration and consumption. This would be used in multi-select scenarios where we would like to highlight several meshes on click and would determine whether or not the highlight layer should be modified on a de-select.