Add muted setting to VideoTexture, fix autoplay in Chrome
Auto-playing a VideoTexture in Chrome failed to work. According to their [autoplay policy](https://developers.google.com/web/updates/2017/09/autoplay-policy-changes) muted autoplay is allowed, so this adds support for a `muted` setting.

But this still seems to be not enough for Chrome, as a video element detached from DOM seems still to not autoplay. See this reproduction: https://jsfiddle.net/6105nLzr/4/ - when line 13 is commented out, the video does *not* start to play (see console output). So calling `play()` programmatically.


Former-commit-id: c99e7b5734605b1edd9f8ad2c2b69e043bce1d86