Changes to XRAnchor system Closing WebXR Anchor system should provide full abstraction on top of XRAnchor #9258
Promises are now resolved only when Babylon has created an IWebXRAnchor object, which also means that the system started tracking it.
Also - IWebXRAnchor has a new remove function that is nothing more than a proxy to the native anchor's delete method.

There is now no need to interact with the xrAnchor, but the access to the object is still available.


Former-commit-id: 7a57ce32801b4d8f74742d5cef48a9383185e2e6