Fix goToFrame when multiple calls are made
When in a paused state, multiple calls to "gotoFrame" create an incorrect _manualJumpDelay. This occurs because the currentFrame of the first RuntimeAnimation is used to determine the that offset. But the currentframe is set on each call regardless of whether the animatable is paused or not. So it uses the offset of the last call to gotoFrame rather than the currentFrame it is actually paused on. I added a local field to keep track of the true currentframe.