Fix glTFLoader 2.0 when dealing with glTF files that contain no meshes
AFAIK glTF may contain no meshes (and hence no nodes) at all (our use case are material-only assets). But in `_setupData` of the glTF 2.0 loader, `_rootBabylonMesh` is only initialized when nodes exist. If that is not the case, `_getMeshes()` returns `[undefined]`, which later throws e.g. in a `AssetContainer.dispose()` call.