[Build] add timestamp to CDN assets (#12353)
Every time a snapshot is created a timestamp.js file willbe updated. this timestamp will be loaded before all babylon resources, and it is never cached.
The timestamp will be used to load from the CDN, which will cache for longer time with the query string included in the cache, meaning - the resources will be cached until a new timestamp is generated.

This way we will not have to force-refresh anymore after a nightly.