Fix distorted normals in large-radius flat IcoSpheres. (#12789)
Prior to this commit, a subdivided IcoSphere with a large radius would have distorted normals
for the vertices that were in the original set prior to subdivision.

After some troubleshooting, I discovered that this problem was due to scaling at an inappropriate
point in the process of generating the mesh. After removing this inappropriate scaling,
the normals of the vertices now look as expected without distortion.

Non-flat icospheres are unaffected - they still look appropriately round, as they did before.