Fix Sound current time when stop is called while paused
When stop is called while paused, the current time includes the start offset set by the pause function. This is not the case when stop is called while **not** paused. This is confusing. When stop is called, the current time should be the same, whether the sound is paused or not paused.

This change fixes the issue by always resetting the start offset when the sound is stopped, even if the sound is paused.