Animations: storing key frames in typed arrays
Keyframes in Babylon.js are stored with array of objects. Some animations exported with millions of key frames, with the object mode, it would take more memory than stored with typed arrays, like Float32Array.
Without the object per frame, it would be much harder for features like per-frame interpolation and tangent to be implemented, so this could be an optional feature for those suffering too many keyframes without these advanced features.
For AnimationCurveEditor it is hard to fully adapt it to CompactAnimation, maybe the fast way it to convert CompactAnimation to Animation, and replace all possible usage to converted animation.

Not supported yet:
* CUBICSPLINE animation with tangent
* per-frame interpolation or tangent
* serializing GLTF animation with quantization and 4-bytes alignment

Forum link:
https://forum.babylonjs.com/t/animations-storing-key-frames-in-typed-arrays/36566