Performance optimization for vector methods
In vector functions like Vector3.FromArrayToRef and Quaternion.FromArrayToRef, values are assigned to properties like x, y, z, w, which are internally setter functions which sets things like _x and _isDirty, so in FromArrayToRef where all these setters called, the _isDirty is set multiple times, which is possibly optimizable.

It should be possible to directly set internal properties like _x, _y, _z, and _w in FromArrayToRef and set _isDirty to true to be compatible with existing code depending on _isDirty, which should result in better performance.

Forum link:
https://forum.babylonjs.com/t/optimizing-performance-of-fromarraytoref/37513