[Suggestion] Notify added observer if observable was triggered.
This is done to avoid the pattern of adding an observer after checking if the state requires that.
As an example let's take the scene's onReadyObservable which has a simple state - it will only notify once when the scene is ready. If a new observer will be added after the observable was triggered (and has notified existing observers) it will not be possible to trigger new observers that were added to is.
If `notifyIfTriggered` is set to true and the observable has already triggered it will automatically trigger newly added observers whenever they are added.

It's also recommended to add that to onDispose observables, which are the same.
A different solution would be to define an onObserverAdded function, but that would require setting this function to every observable that requires it. Instead - this is a more generic solution.