Relax `AssetContainer` heirarchy check to allow `InstancedMesh` parents
When an `InstancedMesh` is used as a parent, `AssetContainer` logs a warning saying the node heirarchy is invalid. This change fixes the issue by changing the parent's `instanceof` check from the `Mesh` class to the `AbstractMesh` class, which includes the `InstancedMesh` class since it is also derived from `AbstractMesh`.

Related forum post:
https://forum.babylonjs.com/t/assetcontainer-throwing-warnings-with-instancedmesh/48857.

Repro playground:
https://playground.babylonjs.com/#H0ALVT#3.