Limit scope of resuming audio context to when XR enters immersive mode
Note that we wait an arbitrary time of 3 seconds for the device to suspend the audio context after immersive mode is entered. Vision Pro suspends the audio context after approximately 0.5 seconds, so 3 seconds should be plenty of time. The user is unlikely to suspend the audio context manually during this 3 second window.