buffer: copy unaligned data in GetFloatData
In `VertexBuffer.GetFloatData`, if data is not aligned to 4 bytes, the current impl is to align the offset to make code work, but this would make data returned here differ from data passed to gpu, so algos depending on result of it could be wrong.
This commit changes this behavior to copying the buffer on unaligned buffers or offsets.

Related playground: <https://playground.babylonjs.com/#5T7E1W#8>
Forum thread: <https://forum.babylonjs.com/t/picked-point-at-wrong-position/51698>