Use gain node for main output's input node so its output can be captured
To capture realtime output from an AudioContext we need to connect the main output's input node to a media stream destination node, but the main output's input node is currently set to the audio context's destination node, which does not allow connecting additional downstream audio nodes.

This change uses a gain node for the main output's input node instead of the audio context's destination node, so additional downstream audio nodes can be connected. This is needed for capturing the audio from streaming sounds for testing since they don't work with offline audio contexts.