Fix regression in static sound `stop` when `waitTime` arg is given
If the static sound `play` function is called with no `await`, a following call to `stop` with a non-zero wait time fails because there are no instances to queue the `stop` on, yet.

This change fixes the issue by making static sound instance creation synchronous instead of asynchronous, which is fine to do since the sound's buffer is populated when `init` resolves.