Fix `WebAudioStreamingSoundInstance` start offset issue
When a streaming instance is paused before it transitions from `Starting` to `Started`, the start offset is being set incorrectly when resumed. This change fixes the issue by setting the related timing variables correctly and initializing the start time to `Infinity` to indicate that the instance has not actually started, yet.