Remove streaming sound play and stop `waitTime` parameters
The WebAudio-based streaming sound class's underlying `HTMLMediaElement` play and stop functions don't have `waitTime` parameters like the static sound class's `AudioBufferSourceNode` does. I tried implementing them in the streaming sound class using `setTimeout`, but the timing is too flaky, so this change removes the parameter completely.

Note that the `waitTime` parameter for the static sound classes remains, but it is moved to the end of the parameter list so the streaming and static sound class play signatures are consistent with each other.