MaterialPluginBase: add doNotSerialize
This commit adds a new public property `doNotSerialize` to `MaterialPluginBase`, so that a material plugin can be configured to skip serialization like meshes and materials.

## Compatibility

Existing material plugins would implicitly have this field as they extend MaterialPluginBase.
Material plugins that does not extend MaterialPluginBase will not have a doNotSerialize after this, but when using if to check, it would have the same result as false, which should be the default value, so backward compatibility for js projects would not be broken after this.
 In typescript projects, type checking ensures that plugins extending MaterialPluginBase.

 ## References

 Forum post: <https://forum.babylonjs.com/t/donotserialize-for-material-plugins/55228>