If HTTP connection closed prematurely, consider retry (#16025)
* If HTTP connection closed prematurely, consider retry

xhr2 (used for HTTP under node) will return response=null if the server
returns a response but the connection is closed before the full
response has been downloaded. Previously this would lead to a null
reference in some handlers, e.g. glTFFileLoader would throw "Cannot
read properties of null (reading 'byteLength')". With this change
onSuccess is not called if the response is null; instead the retry
logic can choose to retry or error.

* Apply formatting changes using prettier