Check if TransformNode is parent in removeChild (#16041)
The function's description is misleading. "Removes the passed mesh from the current mesh children list" does not mean removing the parent from the passed mesh.

This change updates the function to remove the parent from the passed mesh only if it is the parent of the passed mesh. Otherwise, the parent of the passed mesh is left alone.