Skip legacy audio engine init in playgrounds that explicitly turn it off (#16273)
There is currently no way to disable the old audio engine in playgrounds that include `BABYLON.Sound`, even if the playground supplies a `createEngine` function and explicitly sets the `audioEngine` option to `false`.

This change fixes the issue by only doing the backcompat init of the old audio engine if the engine's `audioEngine` option is undefined or explicitly set to `true`. This causes backcompat init of the old audio engine to be skipped when the engine's `audioEngine` option is explicitly set to `false`.