Allow old audio engine to be enabled using engine's `audioEngine` option (#16278)
The old audio engine is not being enabled when the graphics engine is constructed with the `audioEngine` option set to `true`. This change fixes it while keeping the desired behavior of not enabling the old audio engine by default anymore, which was done to pave the way for the new audio engine.