Get old audio engine working in playgrounds that using `BABYLON.Sound`
The previous change to the `audioEngine` option that changes its default from `true` to `false` broke the playground code that initializes the old audio engine when `BABYLON.Sound` is found in the playground's user code.

This change fixes the issue by leaving the `audioEngine` undefined if it is not defined in the playground's user code. This gets the playground's old audio engine init code working correctly without breaking the edge cases fixed in similar previous PRs.