Make sound `pitch` and `playbackRate` properties change continuously
Users expect the `StaticSound.pitch` and `playback` properties to update while a sound is playing, but they don't. They're only updated when playback starts.

This change addresses user needs by updating the pitch and playback rate of the newest sound playback instance immediately instead of only when the `play` function is called.