Skip audio engine v2 auto resume after `pause()` (#16349)
When the audio engine v2 audio context is started and paused later by the browser, the auto resume feature is intended to unlock the audio context automatically as soon as possible. This feature is kicking in when the user explicitly calls the audio engine's `pause()` function, too, though, which is not intended.

This change fixes the audio engine v2 auto resume feature so it doesn't apply when the audio engine's `pause()` function is called.