Remove wait ring at the end of code run process (#16374)
The initial wait ring of the playground is being removed when the code
starts executing. This is an issue for TS-based playgrounds because of
the language initialization.

This removes the wait ring at the end of the load process or when an
error is triggered. This will make sure the wait ring is there until it
is no longer needed.