Fix Viewer toolbar width regression (#16383)
A recent change resulted in the Viewer toolbar having a min-width set
even when there is no animation. In this case, the tool-bar should be
allowed to be narrower since it may only have a single reset button.

This change also brings over two fixes that were included in
https://github.com/BabylonJS/Babylon.js/pull/16372, which was rolled
back:
- Explicitly set the width of the Configurator app level div to 100%.
- Disable/enable snapshot rendering on suspend/resume. This ensures we
can correctly render the final high res frame (before suspend completes)
within a single frame (e.g. single call to scene.render). It will be
slightly slower since it is not using snapshot rendering, but this is
fine for rendering a single frame (fast snapshot rendering is mostly
useful for continuous rendering).