Fix TextLineComponent imports to work outside of the repo (#16387)
When importing @babylonjs/shared-ui-components from another repo, the
build is failing because textLineComponent.js is referencing imports in
this form:
import copyIcon from "shared-ui-components/imgs/copy.svg";

This works in the Babylon repo because that name is mapped to the right
path, but doesn't work when used in another repo. This change fixes the
imports to be relative paths instead.