Change static sound `playbackRate` and `pitch` settings API (#16396)
The static sound `playbackRate` and `pitch` options are confusing when
it comes to sound instances. Changing those options on a sound with
multiple instances playing only updates the most recent instance, but it
should update all instances.

Similarly, allowing the `playbackRate` and `pitch` options to be set in
the `play` function is confusing because the parent sound's properties
will override the new instance's option set by the `play` function,
which will cause the playback rate and pitch to change unexpectedly when
the sound's properties are changed.

This change fixes these issues by making the static sound's
`playbackRate` and `pitch` properties apply to all playing instances,
and by removing the `playbackRate` and `pitch` options from the `play`
function.