Improve ArcRotateCamera interpolation logic (#16430)
This PR makes several improvements to the existing ArcRotateCamera
interpolation logic:
- Allows overlapping interpolations. For example, if you start an
interpolation of alpha, and before it is finished start an interpolation
of the y component of the target, it previously would always stop the
previous interpolation. Now it is possible for them to overlap, but
passing in `NaN` for the components you don't want to interpolate. This
is perhaps a little unusual, but it's the least disruptive to the API.
Still, open to alternate suggestions.
- Ensures the camera orbit/target/etc. eventually reach the actual
interpolation goal values. Previously, it would stop the interpolation
when it was "close," but not actually bring the values to the final goal
values.
- Because of the previous point, I had to improve the logic for
determining when we are "close" to the final goal values, to ensure
"close" is visually indistinguishable from the final value.
- For target, it is relative to the radius. This ensures it visually
always looks the same regardless of scale.
- For rotation, it is just an absolute value (0.0002 radians), which was
chosen empirically by just seeing what looked visually
indistinguishable.
- For radius, it is when we are 99.9% of the way to the goal, which
again should work well regardless of scale.
- For target screen offset, it is just an absolute value, same as
previous actually.
- Swaps the order to applying user input and applying interpolation.
This way, if there is any user input, we don't add in a small and
unnecessary and incorrect interpolation first (size of the change
depends on frame time). Also, it means we don't need to repeat the
checks for user input twice.
- Add a `stopInterpolation` function to make it easier to interrupt an
in-progress interpolation.