Fix lazy loaded property values in audio engine (#16435)
Asynchronously loaded subproperties like sound and bus `analyzer`,
`spatial` and `stereo`, do not return the same values set on them until
the async load completes. This makes it so a property's getter does not
immediately return the value given to its setter. For example
`analyzer.fftSize = 128; console.log(analyzer.fftSize);` prints the
default `2048` until the async `analyzer` subproperty init is complete.
This is confusing for users.

This change fixes the issue by using backing variables for properties on
lazy-loaded sound and bus subproperties.