Viewer: More options + reset logic in Viewer layer (#16436)
This PR aims to solve a couple of different but overlapping problems:
- More flexibility in the reset logic (forum issue:
https://forum.babylonjs.com/t/viewer-v2-reset-button-to-the-initial-explicitly-specified-camera-pose/56925)
- Make it possible to initialize the Viewer from JSON (allows the Viewer
Configurator to be used with all layers of the Viewer, not just the HTML
layer, and also creates a path for saving/loading to/from snippet
server)

Changes to the Configurator will come in a subsequent PR. For this PR,
the overall changes are as follows:
## Viewer
- Flesh out `ViewerOptions` with *most* properties of the Viewer.
- Expose a `DefaultViewerOptions` object, which is used for Viewer
initialization and will also be used by the Configurator.
- Add a `reset` function that helps reset Viewer state to the initial
state (based on passed in options and default options).
- Use this reset state and logic in the Viewer layer and remove old
`UpdateModelOptions` that was needed to coordinate reset logic that was
spread across `Viewer` and `ViewerElement`.
- Update the `resetCamera` function to be more configurable (reframe the
camera (current logic), reset the camera to the initial state, or pick a
reasonable default based on Viewer state).
- Add an `updateCamera` function that will interpolate to a new pose.
- Remove `FramingBehavior`. It was used in early versions of the Viewer,
but at this point the Viewer's framing logic has superseded
`FramingBehavior` and it is no longer needed.
- Move all hotspot state and logic down to the Viewer layer.

## ViewerFactory
- Wrap the passed through options in a `Proxy`, which makes is possible
for the higher layers (e.g. `ViewerElement`) to override the default
options dynamically (used to allow element attributes to behave as
default options).

## ViewerElement
- Initialize properties/attributes based on the passed in options (since
they now have most of the Viewer properties).
- Pass `ViewerOptions` down that return attributes if they are set,
otherwise the passed in option value. This makes it so if an attribute
is dynamically updated, it becomes the configured value that is used by
the reset logic.
- Change `camera-orbit` and `camera-target` to be custom handled
attributes (not reactive properties). There is a lot of special handling
to get the desired behavior for camera pose, so the reactive property
pattern is not helpful anymore.
- Add a `resetMode` property/attribute, and a `reset` function. The
behavior of the `reset` function (called when the reset button is
clicked) depends on the `resetMode`. This is what finally addresses the
forum issue.
- Remove `syncToAttribute` logic that was previously used for reset
logic (since reset logic moved down to the Viewer layer).
- Remove logic to reset camera state, since it also moved down to the
Viewer layer.
- Remove `UpdateModelOptions` related args previously passed to
`Viewer.loadModel` as this is now all handled in the Viewer layer.
- Remove all hotspot implementation since it is now handled in the
Viewer layer.