Viewer Configurator: Add json output + save/load snippet (#16447)
This change introduces two new configurator features:
1. The ability to save/load the current state (to/from snippet server).
See https://forum.babylonjs.com/t/babylon-viewer-v2/54317/65

![image](https://github.com/user-attachments/assets/bd33db53-3c56-4524-ac07-2be32947b4f8)

![image](https://github.com/user-attachments/assets/116bc8c6-d014-47cd-aa7e-7dd52e0c2dd8)

1. The ability to output to json, which can in turn be passed into any
of the Viewer layers as the `ViewerOptions` (making the configurator
useful even if you aren't using the html layer).

![image](https://github.com/user-attachments/assets/0a12b30a-36d6-428b-95fe-590f7219dffe)

Changes to enable this:
- Add format dropdown to choose between html and json (html by default).
- Improve camera pose management.
- Pass an additional `initialConfiguredState` to `useConfiguration`
(comes from the loaded config/options).
- Use loaded options and default options in all calls to
`useConfiguration`.
- Add the ability to save the current configuration (as `ViewerOptions`)
to snippet server and encode into the url like other Babylon tools.
- Add the ability to load a configuration from snippet server (via url)
and use it with `ConfigureCustomViewerElement` to determine the initial
state of the Viewer within the Configurator.

Also fixed a small bug where environment intensity and rotation were not
output if the skybox was disabled.