glTF Exporter: Fix LH -> RH vertex data conversion bug (#16456)
When iterating over left-handed vertex buffers to perform our RH
conversions, we can accidentally iterate several times over the same
buffer if it is used by multiple meshes. See
https://playground.babylonjs.com/#KX53VK#85 and update `meshCount` for
demo of issue.

- Calculate the `maxTotalVertices` to use when enumerating over the
vertex data, which is the largest number of vertices in any mesh using
the buffer. This should cover any cases where meshes use only a
subsection of the vertex buffer, starting from its beginning. There
should not be any cases where the `maxTotalVertices` < the union of the
ranges of the buffer that are used by individual meshes.
- Factor out common vertex buffer info retrieval into utility function,
`GetVertexBufferInfo`
- Simple visual test based on playground from above. (Considered making
this a unit test, but it'd be lengthy, as we're testing a roundtrip +
we'd need to parse the buffers.)
- Simple unit testing for converting float matrices indices