glTF Loader: Get `mimeType` from URI if not defined (#16462)
glTF images that did not specify a `mimeType` would not set the
`Texture.mimeType` field. This happens with GLBs, for example. While the
field isn't required, it helps with export. The glTF exporter uses this
field to try to preserve the original mime type when re-encoding the
images-- if undefined, they'll default to PNG. The Babylon exporter
should also be using this information when serializing its image
buffers.