Add empirical modification to specular reflectance to handle low IOR's (#16452)
Babylon uses the split sum approximation to apply the BRDF and this
leads to an issue when IOR gets close to 1.0. When inside and outside
IOR's match, specular reflectance should go to zero. However, because
the BRDF texture is baked for an IOR of 1.5 and we only use IOR to
modify F0, the specular reflectance at glancing angles ends up very
wrong when IOR gets close to 1.0.

I've added a linear scale to the F90 value to help account for this.
It's still not correct, of course. We'd need to do the actual Fresnel
calculations for the given IOR to fully fix the problem (note that the
Schlick approximation is not sufficient).

IBL:

![ReflectanceIOR](https://github.com/user-attachments/assets/cc24a000-fde9-4e03-9229-587d7cf06e69)

Analytic Lights

![ReflectanceIOR_analytic](https://github.com/user-attachments/assets/212383f4-8178-48a9-9087-06c7b8815f57)