Interactivity fixes (#16473)
A few fixes after testing some more demo files:

1. Extension path matching is optional in certain cases. So even if a
value doesn't exist in the glTF tree itself, we should still be able to
get and set it. Good examples are selectability and visibility - both
exist on the babylon object, but are optional in the glTF tree.
2. Parsing NaN and Infinity now works (if they are provided as strings)
3. Interpolation block was not providing the context to the animation
function
4. Matcher for json pointer didn't work correctly if the matched item
was at the end of the string
5. pointer/interpolate should use the BezierEasing and not the normal
easing block
6. Small fix for out flow of the send event block.