Shader pre-processor: Improve parsing + fix WGSL (#16474)
I have improved the shader pre-processor so that constructions of this
type are analyzed correctly:
```glsl
#define BRDF_DIFFUSE_ROUGHNESS_EON 0
#define BRDF_DIFFUSE_ROUGHNESS_BURLEY 1

#if BASE_DIFFUSE_ROUGHNESS_MODEL == BRDF_DIFFUSE_ROUGHNESS_EON
    diffuseRoughnessTerm = diffuseBRDF_EON(vec3(1.0), diffuseRoughness, NoL, NoV, LoV) * PI;
#elif BASE_DIFFUSE_ROUGHNESS_MODEL == BRDF_DIFFUSE_ROUGHNESS_BURLEY
    diffuseRoughnessTerm = vec3(diffuseBRDF_Burley(NoL, NoV, VoH, diffuseRoughness) * PI);
#endif
```
Before the PR, you couldn't test against a name, you had to use
immediate values (0 instead of `BRDF_DIFFUSE_ROUGHNESS_EON` and 1
instead of `BRDF_DIFFUSE_ROUGHNESS_BURLEY` in the example above).

I have also reworked the way we handle preprocessing in WGSL, as it was
causing some problems before. I have tried to add a number of comments
in the code to explain why things are done in that way.