Fix flaky playwright audio tests (#16478)
The playwright audio tests expecting volume curve values are flaky
because the volume curve analysis algorithm is leaving a zero volume gap
sample between each pulse. If the time the volume is checked at falls on
that gap, the test fails.

This change fixes the issue by removing the gap in the volume curve and
writing the average volume sample values across the entire span of the
pulse without skipping a sample between pulses.