Fix for null TypeError in dynamicTexture (#16491)
It's possible that when rapidly disposing a dynamicTexture, a delayed
callback gets ran on the disposed instance, resulting in a TypeError
exception caused by accessing a null instance. The specific case that
was causing this issue was when `Tools.SetImmediate` is used to pend
some future work to `update` the dynamic texture, and by the time this
delayed callback was running, the dynamicTexture had already been
disposed.

The fix here is to early out of the `update` function if `this._texture`
is already null. This simply avoids dereferencing a null engine when
calling `_getEngine().updateDynamicTexture(this._texture, ...)`. In any
case, `updateDynamicTexture` itself already early outs when the texture
parameter is null, so it should not change that behavior.