Viewer: Remove unused UV sets (#16495)
This change is in response to feedback from users that certain models
don't render in the Viewer (@alexandremottet). For the model in
question, the problem is that it has a bunch of unused uv sets in the
vertex data, which exceed the WebGPU limits, and WebGPU is the default
for the Viewer in many cases. The fix is to add a utility function that
removes uv sets that are in the vertex data but unreferenced by any
material.
- Add `removeVerticesData` to `AbstractMesh`, and add an implementation
in `InstancedMesh`.
- Add a `RemoveUnreferencedVerticesData` helper function to core.
- Call `RemoveUnreferencedVerticesData` after loading a model in the
Viewer.

I also included a few other small fixes in this PR:
- Fix the naming of `CreateViewerForCanvas` (it was inadvertently
exported with the wrong casing). This is a breaking change, but the API
is marked as `@experimental`, so I think we should go ahead and fix it
and if we get asks, we can always export the old casing and mark it
deprecated.
- Export `ConfigureCustomViewerElement` (it was missing from the top
level exports, but is part of the public API).
- Expose a `ViewerElement.reload` function, which is needed when
replacing the `reload-button` slot (previously it called a private
function, which external code could not do).
- Remove the `animationAutoPlay` option from the Configurator's use of
`LoadModelOptions` as this option no longer exists.