Speed up playwright audio tests (#16496)
This change does 3 things to speed up the playwright audio tests:

1. Loads only the bare minimum html needed to run the tests:
- The tests are currently loading empty.html from the CDN, which
downloads a lot more than the audio tests need.
- This change loads custom html with an empty body and 2 scripts from
the CDN; babylon.js and audiov2-test.js.

2. Uses the WebAudio API `OfflineAudioContext` in place of the realtime
`AudioContext` where possible.
- The `OfflineAudioContext` renders audio faster than realtime and can
be used for most of the audio tests. Notable exceptions are the
streaming sound tests and the current time tests.

3. Passes less test result info from the playwright browser back to
Node:
- The tests are currently passing the captured audio to Node after every
test, which is expensive.
- This change does the required audio processing in the playwright
browser and only passes the result back to Node, which is much smaller
than the entire captured audio array.

This takes the AudioEngineCI pipeline "run tests" step from 38 minutes
to 12.