Fix sandbox animation issue
The sandbox errors out when the following steps occur:
1. Load a model with multiple animations.
2. Select the 2nd animation, or any animation other than the 1st.
3. Load a model with one animation.

The animation group index set on the 1st model does not get updated when a new scene is loaded which makes the animation renderer index into a non-existent animation group after the 2nd model is loaded.

This change fixes the issue by resetting the animation state's `groupIndex` to `0` when a new model/scene is loaded.